First official commit of the wand carbon project.
Doesn't yet compile to anything useful.

Added a trait TransferComponent which modules can extend to contribute to the
transfer operation.

Added a WandModule which should be responsible for all wand related things.
Right now it has the component registry for the transfer components, but maybe it
might be cleaner to move this to a separate TransferModule.

Changed the StateModule to be able to create a completely fresh empty state.
Maybe the code to actually "havoc" variables should go into the state components.