Well definedness checking order fixes (#429)
Fix well-definedness checking order in various cases (PR #429)

For various expressions, Carbon checks well-definedness in an unintuitive order that is not in-sync with Silicon. This pull request changes the order of checks in the following cases (the descriptions provide the new checking order):

 - Division a/b: Definedness of a and b are checked before checking that b is non-zero.
 - Modulo a mod b: Definedness of a and b are checked before checking that b is non-zero.
 - Map and sequence lookup: Keys and indices are checked first before checking whether the key/index is in the sequence/map

To simplify this change the method "simplePartialCheckDefinedness" in the DefinednessComponent has been split into two methods (one for checks before checking subnodes, which is the same as the one earlier and one for checks after checking subnodes).