Quasihavoc initial support (#439)
This PR adds initial support for quasihavoc statements to Carbon (see the corresponding Silver PR https://github.com/viperproject/silver/pull/612). Quasihavocs for fields and predicates are desugared into exhaling all the permission of the corresponding resource and inhaling the same permission. For all other quasihavocs, consistency errors are reported (this is global, if there is one such unsupported quasihavoc no method is verified; this is the main reason for using consistency errors).