Refactor well-definedness of field acesses and permission division
Now, the well-definedness of a field access first checks the well-definedness of a receiver and only afterwards checks whether there is enough permission (before the order was reversed).
Now, the well-definedness of a permission division first checks the well-definedness of the dividend and the divisor and only afterwards checks the whether the divisor is nonzero (before the order was reversed).

Moreover, in order to avoid checking whether permission to field accesses at the top-level of an accessibility predicate or permission introspection, the current implementation uses global state in a somewhat fragile way. Now, a different approach is used that avoids global state.